requirements.txt
Pillow>=6.2.0
python-twitter>=3.5
python-dotenv>=0.10.3
